server/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Environment & Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
